+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchModule.kt
@@ -4,6 +4,7 @@ import com.f2prateek.rx.preferences2.Preference
 import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
+import org.simple.clinic.remoteconfig.ConfigReader
 
 @Module
 object OverdueSearchModule {
@@ -13,5 +14,12 @@ object OverdueSearchModule {
       rxSharedPreferences: RxSharedPreferences
   ): Preference<Set<String>> {
     return rxSharedPreferences.getStringSet("preference_overdue_search_history_v1")
+  }
+
+  @Provides
+  fun overdueSearchConfig(
+      configReader: ConfigReader
+  ): OverdueSearchConfig {
+    return OverdueSearchConfig.read(configReader)
   }
 }